+++ variables.tf
@@ -227,6 +227,11 @@ variable "ingress_custom_configuration" {
   default = null
 }
 
+variable "ingress_extra_args" {
+  type    = map(string)
+  default = {}
+}
+
 # ================== ingress-nginx-additional =================
 
 variable "helm_ingress_nginx_additional_enabled" {
@@ -298,6 +303,11 @@ variable "cert_manager_chart_version" {
 
 variable "cert_manager_priority_class_name" {
     default = ""
+}
+
+variable "ingress_additional_extra_args" {
+  type    = map(string)
+  default = {}
 }
 
 # ================== loki-distributed ================= #
@@ -571,7 +581,7 @@ variable "helm_prometheus_enabled" {
 }
 
 variable "prometheus_chart_version" {
-    default = "35.0.3"  
+    default = "44.2.1"  
 }
 
 variable "prometheus_priority_class_name" {